+++ app/upgrades.go
@@ -21,7 +21,7 @@ import (
 	upgradetypes "github.com/cosmos/cosmos-sdk/x/upgrade/types"
 )
 
-func (app *mrmintchainApp) RegisterUpgradeHandlers() {
+func (app *MrmintchainApp) RegisterUpgradeHandlers() {
 	planName := "integration-test-upgrade"
 	app.UpgradeKeeper.SetUpgradeHandler(planName, func(ctx sdk.Context, plan upgradetypes.Plan, fromVM module.VersionMap) (module.VersionMap, error) {
 		return app.mm.RunMigrations(ctx, app.configurator, fromVM)