+++ testutil/tx/cosmos.go
@@ -57,7 +57,7 @@ type CosmosTxArgs struct {
 // It returns the signed transaction and an error
 func PrepareCosmosTx(
 	ctx sdk.Context,
-	appmrmintchain *app.mrmintchainApp,
+	appmrmintchain *app.MrmintchainApp,
 	args CosmosTxArgs,
 ) (authsigning.Tx, error) {
 	txBuilder := args.TxCfg.NewTxBuilder()
@@ -90,7 +90,7 @@ func PrepareCosmosTx(
 // the provided private key
 func signCosmosTx(
 	ctx sdk.Context,
-	appmrmintchain *app.mrmintchainApp,
+	appmrmintchain *app.MrmintchainApp,
 	args CosmosTxArgs,
 	txBuilder client.TxBuilder,
 ) (authsigning.Tx, error) {