+++ x/feemarket/keeper/keeper_test.go
@@ -44,7 +44,7 @@ type KeeperTestSuite struct {
 	suite.Suite
 
 	ctx         sdk.Context
-	app         *app.mrmintchainApp
+	app         *app.MrmintchainApp
 	queryClient types.QueryClient
 	address     common.Address
 	consAddress sdk.ConsAddress